use single command to uninstall mesh and cluster wide resources
This commit combines the osm uninstall mesh and osm uninstall
cluster-wide-resources to a single command. The user can specify
delteion of cluster wide resources by using the flag
--delete-cluster-wide-resoures.

If there is no mesh in the cluster, the user can still use the same
command to delete the cluster wide resources by providing the flag.

 This command has been made idempotent as
well.

fixes #4429

Signed-off-by: Sneha Chhabria <snchh@microsoft.com>