fix(cli): uninstall mesh force flag description
When `osm uninstall mesh -f` is run, the `force`
flag is intended for the user to skip the uninstall
confirmation prompt only.

Even if this code is specified, the code in uninstall
mesh does not supress diagnostic messages or errors
when it encounters an error during mesh uninstallation.

This proposed fix changes the flag description to make
this intended behavior clearer.

Signed-off-by: Johnson Shi <Johnson.Shi@microsoft.com>