feat(cli): add support controller-debug-server cmd
Adds "osm support controller-debug-server" cmd,
which performs port-redirection and opens the browser
to facilitate opening the osm-controller debug server.

This is useful for both project maintainers and osm
users that want to debug a running osm instance without
needing to clone the repo to run
"scripts/port-forward-osm-debug.sh" or to remember
kubectl port-forward commands.

Signed-off-by: Johnson Shi <Johnson.Shi@microsoft.com>