ref(build): update image scanning
This change cleans up the Makefile so the `trivy-scan-images` target no
longer enumerates each image to scan and uses `OSM_TARGETS` instead so
those commands don't need to be updated when images are added or
removed. This change also allows `make trivy-scan-images` to be run
locally when `trivy` exists on `$PATH` when the 64-bit Linux binary
downloaded for CI is not compatible.

The version of `trivy` used in CI has also been bumped from v0.18.0 to
v0.23.0.

Signed-off-by: Jon Huhn <johuhn@microsoft.com>