ref(install): remove redundant checks (#4543)
This change removes 2 checks from `osm install` that are already
performed by Helm or Kubernetes:

The check that the mesh name is valid is done by Helm first when it
validates the release name. Helm release names have stricter
requirements than Kubernetes label values, including a shorter max
length (53) and no underscores allowed. Even if it becomes possible in
the future for a valid release name to be an invalid label value,
Kubernetes will validate the mesh name as a label value when the
osm-controller deployment is created, since it uses the mesh name as a
label value. Validating this was necessary before the mesh name label
was added to make sure other commands like `osm namespace add` would
work once the mesh is installed.

The check that the mesh name is unique is done by Helm when it verifies
no resources from the chart already exist. Since the chart names the
cluster-wide RBAC resources with the mesh name, Helm by itself will fail
when installing a mesh with the same name as an existing one.

Fixes #2147

Signed-off-by: Jon Huhn <johuhn@microsoft.com>