Create release process documentation (#4557)
It is important that as an open source software project we have clarity
and transparency on how releases are handled. This includes the
frequency as well as how the community can request for changes. This
also goes into depth about the lifecycle of a request for change and how
it makes it into a release.

Signed-off-by: Thomas Stringer <thomas@trstringer.com>