fix(healthprobes): add support for TCPSocket probes (#4558)
Currently, OSM rewrites the port value on TCPSocket probes
to the osm reserved port for the respective probe type. An envoy
listener and cluster are configured at the rewritten port.
When performing the health probe, the kubelet will attempt to
open a socket at the specified port. In this case, the specified
port is open on the envoy container. Therefore, the kubelet will
consider the container healthy and the probe will succeed even
though it never checked if a connection could be established at
the original port. False positive TCP connections are a known
limitation of Envoy.

This solution transforms TCPSocket probes on injection into
HTTPGet probes. Rewrites the probe's port to a new osm reserved
port for the osm-healthcheck HTTP server and updates the path to
/osm-healthcheck. A header `Original-Tcp-Port` is also added to
the new HTTPGet probe and set to the port specified on the
original TCPSocket probe.

osm-healthcheck will handle HTTP health probe requests from
kubelet, get the original TCP port from the request's headers
and attempt to open a socket at the specified port. The response
status code corresponds to the success or failure of establishing
a connection.

Signed-off-by: jaellio <jaellio@microsoft.com>