crd/conversion: only patch CRDs needing conversion
Updates the code to only patch CRDs that require
a conversion strategy. This prevents the webhook
from being invoked for resources that do not
require conversion, thereby reducing the load
on the webhook server in large scale environments.
It also removes an additional failure path for
resource CRUDs.

Note: the existing stubs for APIs that do not need
a conversion are still present, in case they need
to be reused later on.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>