feat(certificates) begin to abstract the cert manager patterns
This change pulls out the shared logic between the cert providers into
a new manager struct. This is done in pieces, to avoid a single large PR
with O(1000)'s of LOC changed. The goal is to reduce all of the existing
"providers" to clients that only deal with certificates (no caching, msg broker,
logging, or timing).

Once all of these are gone, we will remove the certificate.Manager interface
and promote the certificate.manager struct to the exported certificate.Manager
(struct).

Signed-off-by: Sean Teeling <seanteeling@microsoft.com>