ref(webhook): remove leaf cert from webhook cabundles (#4603)
Currently, the CABundles in the Webhook Configuration are set to
the cert chains of the webhook certificates. The cert chains
include the root, intermediate, and leaf certificates and
the cabundle only typically includes the root and intermediate
certificates.

As described in #2650, Kubernetes secrets for the webhook certs
were used to fix a race condition on webhook certificate creation.
Since the CABundles included the leaf certificate and each webhook
had its own cert, the Kubernetes API would only be able to
successfully validate one webhook server instance of each type.

This change sets the CABundle for each Webhook Configuration to be
the IssuingCA rather than the CertChain. As a result, the secrets
for the webhook certificates are no longer necessary.

Signed-off-by: jaellio <jaellio@microsoft.com>