feat(metrics): add osm_events_queued metric (#4670)
This change adds an `osm_events_queued` metric which tracks how many
events have been placed on the message broker's workqueue that have not
yet been processed.

Example:

    osm_events_queued 0

The metric is read from polling the workqueue's `.Len()` since it does
some accounting behind the scenes to combine equivalent items, so the
number of items placed on the queue is not necessarily equal to the
number of items taken off the queue when the queue is empty. The polling
interval is set to 5s because there is a visible impact to CPU when this
is as low as 1s and 5s is frequent enough to update at least once
between each Prometheus scrape every 10s by default.

Rate limited events are only placed on the workqueue after some delay
defined by the rate limiter. Until then, those events are not included.

Part of #4568

Signed-off-by: Jon Huhn <johuhn@microsoft.com>