ref(injector): load bootstrap SDS configuration from filesystem (#4635)
Loads the SDS configuration for the envoy bootstrap config 
from the file system. Previously, the certificates were provided
inline to the Envoy bootstrap configuration which was stored
in k8s Secret volume mounted to the Pod. This change will 
allow OSM to update the certificates used when establishing a
connection with xDS without having to recreate the Pod. This
allows the envoy xDS certificates to be rotated without potential
data plane downtime. 

This change updates the existing envoy bootstrap k8s secret
to include the TLS and validation context SDS configs, xDS 
certificate, key, and ca cert.

This change is a part of the automated root certificate rotation
work (#4502)

Signed-off-by: jaellio <jaellio@microsoft.com>