feat(injector): add list of ignored network interfaces (#4700)
* feat(injector): add list of ignored network interfaces

This change adds a new configurable list of ignored network interface
names. Traffic received from and sent to those interfaces is not
forwarded to the sidecar container by iptables. When this list is empty
(the default), osm-injector produces the exact same iptables commands as
it did before.

The list is configured in the chart at `osm.networkInterfaceExclusionList`
and in the MeshConfig at `spec.traffic.networkInterfaceExclusionList`.

Fixes #4546

Signed-off-by: Jon Huhn <johuhn@microsoft.com>

* codegen

Signed-off-by: Jon Huhn <johuhn@microsoft.com>

* add comment

Signed-off-by: Jon Huhn <johuhn@microsoft.com>
Co-authored-by: Shashank Ram <21697719+shashankram@users.noreply.github.com>