Introduce service.ProviderMapper to MeshService
As a stepping stone for statefulset support (#3477), introduce a new
interface describing the ability to map an entity back to a provider
service (e.g. a Kubernetes service). This decouples the MeshService name
from being a foreign key between the provider's collection of services
and the MeshCatalog's set of services

Signed-off-by: Keith Mattix II <keithmattix2@gmail.com>