feat(certs): create MRC on install
Updates the osm-bootstrap to create a default MeshRootCertificate
on osm install. Adds a preset-mesh-root-certificate ConfigMap to
the Helm templates. The osm-bootstrap will obtain the MRC spec
from the ConfigMap and attempt to create a default MRC. If an MRC
already exists with complete state and issuing rotationStage then
the osm-bootstrap will not create the default MRC.

Additional changes:
- adds MeshRootCertificate CRD to uninstall list
- adds a tokenSecretName value to the osm.vault values

Signed-off-by: jaellio <jaellio@microsoft.com>