cli/verifier: add control-plane-health command
Adds a `verify control-plane-health` command.
Currently this check is limited to verifying
that the control plane pods are running.
This will be extended to perform additional
checks pertaining to the health of the
control plane, such as looking for fatal/error
events reported, probe failures, etc.

Also fixes a linting issue in egress test.

Part of #4634

Signed-off-by: Shashank Ram <shashr2204@gmail.com>