feat(certificate): reduce methods on certmanager.
There are 2 exported methods on certmanager that are only used in unit
tests in other packages. This removes that need by leveraging interfaces
in the appropriate places where a test may be needed. The 2 methods are
GetRootCertificate and GetCertificate. This is especially useful for both
upcoming changes, and when installing certmanager, as it removes the burdensome
step of needing to place the Root CA in a k8s secret

Signed-off-by: Sean Teeling <seanteeling@microsoft.com>