cli/verifier: use pod status conditions for readiness check (#4749)
The pod status `phase` does not accurately convey whether
a pod is ready to serve requests, as intended by the
verifier. The `running` phase only implies:
"The Pod has been bound to a node, and all of the
containers have been created. At least one container
is still running, or is in the process of starting or
restarting."

This change updates the verifier to examine the
more granular pod status conditions.

Part of #4634

Signed-off-by: Shashank Ram <shashr2204@gmail.com>