ref(certs): use secretKeyRef for Vault token in MRC (#4736)
Updates the MeshRootCertifcate CRD to use secretRefs for the
Tresor CA and a secretKeyRef for the vault token. Currently, the
Vault token would be set as a string in the MeshRootCertificate.
The token needs to be stored securely and not inline in the MRC.

This change also implies that users will be able to reference secrets
containing the CA for Tresor in any namespace. Currently, the
Kubernetes secret (by default called osm-ca-bundle) is assumed
to be in the OSM control plane namespace. If we want to enforce
this, we could validate the MRC or only allow the secretRef's
name to be set and the namespace could be populated by OSM.

Signed-off-by: jaellio <jaellio@microsoft.com>