feat(certs): get Vault token from Secret
If the Vault token option is not set, the token will be obtained
from the secret referenced in the SecretKeyRef. Currently, the
Vault token must still be set as an option if Vault is the cert
provider. In a subsquent change, all cert provider flags will be
removed from the osm-bootstrap, osm-controller, and osm-injector.
These values will instead be obtained from the MRC created by the
osm-bootstrap and populated using the preset-mesh-root-certificate

Signed-off-by: jaellio <jaellio@microsoft.com>