feat(certs): create MRC on install (#4747)
Updates the osm-bootstrap to create a default MeshRootCertificate
on osm install. Adds a preset-mesh-root-certificate ConfigMap to
the Helm templates. The osm-bootstrap will obtain the MRC spec
from the ConfigMap and attempt to create a default MRC. If an MRC
already exists with a complete state and issuing rotation stage then
the osm-bootstrap will not create the default MRC.

Additional changes:
- adds MeshRootCertificate CRD to uninstall list
- adds the tokenSecretName and tokenSecretKey values to the
osm.vault values
- removes SecretName from the certManager provider in the mesh
root certificate
- up MeshConfig tests

Part of #4712

Note: This change does not include creating the Vault token
secret. The MRC created is not used at this point by the OSM
control plane. This will come in a later PR.

Signed-off-by: jaellio <jaellio@microsoft.com>