ref(ci): update actions/setup-go to v3
v3 of the setup-go action used in CI has built-in build and module
caching which allows us to remove a good amount of boilerplate. It can
also select a Go version to use based on the `go` directive in go.mod
so updating the version of Go we want to use won't require changing the
CI config at all.

Signed-off-by: Jon Huhn <johuhn@microsoft.com>