ref(ci): run tests/scenarios as unit tests
The tests defined in tests/scenarios included a e2e-like test that
required a running Kubernetes cluster until it was removed in #4154.
Now, all of the tests there can be run just as easily as the rest of the
unit tests. This change will result in tests from tests/scenarios being
run as part of the unittest CI job and removes the special config for
them that's no longer necessary.

Signed-off-by: Jon Huhn <johuhn@microsoft.com>