fix(certs): updates checkAndRotate to use current durations
Updates the checkAndRotate method to update certs with the current
validity duration specified in the MeshConfig for their respective
cert type. Adds a certType to the certificate struct that is used
to lookup validity duration when calling IssuingCertificates.
Updates shouldRotate to check if the cert needs to be rotated due
to an in progress root certificate rotation. GetFromCache will
return nil if the cert is not found in the cache, expiring, or
needs to be rotated because of a root cert rotation.

Signed-off-by: jaellio <jaellio@microsoft.com>