envoy|catalog: use TrafficMatch to build inbound filter config
Similar to how outbound filter config is built, this change
relies on the TrafficMatch type to build inbound filter
configs. The TrafficMatch has a 1-1 mapping to an Envoy
filter_chain config. This is necessary to be able to
cleanly apply TrafficMatch based policies (e.g. rate
limiting) without needing to rely on multiple catalog
APIs.

Part of #2018

Signed-off-by: Shashank Ram <shashr2204@gmail.com>