fix(certs): update checkAndRotate to use current durations (#4800)
Updates the checkAndRotate method to update certs with the current
validity duration specified in the MeshConfig for their respective
cert type. Adds a certType to the certificate struct that is used
to lookup validity duration when calling IssuingCertificates.
Updates shouldRotate to check if the cert needs to be rotated due
to an in-progress root certificate rotation. 

Signed-off-by: jaellio <jaellio@microsoft.com>