Separate bootstrap building logic into the envoy/bootstrap package (#4838)
This PR separates the bootstrap config build logic into the bootstrap
package. There are 0 "logic" code changes, only moving code around,
exporting previously unexported functions, and prefacing 
methods/types with the new package name.

The main reason to do this is that there are changes coming with
respect to rotating the secrets with the bootstrap config. Because
of this, we will need to modify how we are creating and writing the
bootstrap config.

Signed-off-by: Sean Teeling <seanteeling@microsoft.com>