Add GitHub Action to require size and kind labels (#4876)
* Add GitHub Action to require size and kind labels

Currently there is no enforcement of adding `size/*` or `kind/*` labels
to issues. This will help the project and collaborators understand the
type of issue, as well as a rough estimate of how long it is projected
to take to resolve.

This PR adds the workflow to mark issues with `size/needed` in the
absence of `size/*` labels, and `kind/needed` in the absence of `kind/*`
labels. This will only consider issues that are part of a milestone.

Signed-off-by: Thomas Stringer <thomas@trstringer.com>

* Fix typo for configuration setting

Signed-off-by: Thomas Stringer <thomas@trstringer.com>