Collapse informer event handlers into the informer collection.
This contains changes to the events.PubSubMessage to split the type
and kind into the k8s object, and update/delete/added methods.

This allows the informer to dynamically infer each message type.

Signed-off-by: Sean Teeling <seanteeling@microsoft.com>

Setup events for dynamic discovery of the kind

Signed-off-by: Sean Teeling <seanteeling@microsoft.com>

use workqueue for proxy updates

Signed-off-by: Sean Teeling <seanteeling@microsoft.com>

Move most of the configurator methods onto the mesh config object itself

Signed-off-by: Sean Teeling <seanteeling@microsoft.com>

progress on test cases

Signed-off-by: Sean Teeling <seanteeling@microsoft.com>