Remove pre-povisioned ADS certificate (#5101)
Removes the pre-provisioning of the ADS cert that is provisioned. 
When it was pre-provisioned it ended up with the ads.cluster.local 
trust domain appended. When rotating it would end up using this 
even though it was requested that the provisioned cert already 
included the FullCNProvided which would have given a cert with 
CN ads. It was an implementation detail that the cert that was 
pre-provisioned ended up being re-used because our cache 
mechanism looks up the certs by "prefix" which was ads.

Signed-off-by: James Sturtevant <jstur@microsoft.com>