Refactor IssueCertificate API use Options pattern
This does a few things:

- make it possible to add additional information
in the Providers IssueCertificate. For instance,
we re-use the information to generate a SpiffeID,
where as prior we needed guess the trust domain
and deconstruct the CN to figure it out
- Simplifies the calls to the Managers
IssueCertificate API.
In the case of prefix it was sometimes the prefix,
 sometimes not. In all cases it was the cache
key and we can hide that detial further.
- The certtype is internal to manager and will be
removed in future when the information is stored on
the MRC so this is a step in that direction by only
using it in cert package and unexporting it.

Signed-off-by: James Sturtevant <jstur@microsoft.com>