k8s: update app protocol detection for k8s service port
This change updates the app protocol detection for
a service port to eliminate an error scenario that
results from using the same port for multiple protocols.

It does the following:

1. Only uses http implicitly for TCP based ports
   that do not have appProtocol or a valid protocol
   prefix in the port name.

2. Adds a warn log when implicitly using http as
   the app protocol for TCP based ports.

The result is that duplicate filter chain matches
are eliminated with the same port is used for different
protocols, eg. port 8080 being used for TCP and UDP.

Resolves #5223

Signed-off-by: Shashank Ram <shashr2204@gmail.com>