fix: remove Automatic rebase mergify rule
Remove the Automatic rebase mergify rule due to mergify
deprecation of using a random org user as the bot_account.
Mergify instead uses the author of the PR. The action will fail
if the user has not logged into the mergify dashboard. If a
github application such a dependabot creates a PR and triggers the
rebase mergify action, the action will fail.

Signed-off-by: jaellio <jaellio@microsoft.com>